+++ src/audio/playMusic.ts
@@ -8,6 +8,8 @@ export function playMusic(url: string, opt: AudioPlayOpt = {}): AudioPlay {
     const el = new Audio(url);
     el.crossOrigin = "anonymous";
     el.loop = Boolean(opt.loop);
+    el.volume = opt.volume ?? 1;
+    el.playbackRate = opt.speed ?? 1;
 
     const src = _k.audio.ctx.createMediaElementSource(el);
 
@@ -90,11 +92,12 @@ export function playMusic(url: string, opt: AudioPlayOpt = {}): AudioPlay {
         },
 
         set detune(d) {
-            // TODO
+            throw new Error(
+                "Music cannot be detuned (limitation of the browser APIs, womp womp)",
+            );
         },
 
         get detune() {
-            // TODO
             return 0;
         },