+++ src/core/contextType.ts
@@ -1,3 +1,4 @@
+import type { FixedSpeedOption } from "../app/app";
 import type {
     ButtonBinding,
     ButtonBindingDevice,
@@ -162,7 +163,9 @@ import type { Mat4 } from "../math/Mat4";
 import type {
     Circle,
     Ellipse,
+    getSpriteOutline,
     Line,
+    Mat2,
     Mat23,
     Point,
     Polygon,
@@ -173,6 +176,7 @@ import type {
     StepPosition,
 } from "../math/math";
 import type { NavMesh } from "../math/navigationmesh";
+import type { Quadtree } from "../math/spatial/quadtree";
 import type { Vec2 } from "../math/Vec2";
 import {
     type Anchor,
@@ -863,6 +867,35 @@ export interface KAPLAYCtx {
      * @subgroup Rendering
      */
     polygon(pts: Vec2[], opt?: PolygonCompOpt): PolygonComp;
+    /**
+     * Create an outline in the form of a polygon from a sprite object
+     *
+     * @param asset - The name of the sprite you want to generate an outline for
+     * @param frame - The frame of the sprite that the outline should be taken from
+     * @param RDP - Should the function return a simplified polygon
+     * @param epsilon - The amount of the polygon that should be reduced
+     *
+     * @example
+     * ```js
+     * // Create a collider from the sprite
+     * onLoad(() => {
+     *     add([
+     *         sprite("apple"),
+     *         area({shape: getSpriteOutline("apple")}),
+     *         pos(120, 80), // optional position
+     *     ]);
+     * });
+     * ```
+     *
+     * @returns A polygon comp.
+     * // TODO: add since, group and subgroup
+     */
+    getSpriteOutline(
+        asset: string,
+        frame?: number,
+        RDP?: boolean,
+        epsilon?: number,
+    ): Polygon;
     /**
      * Attach and render a rectangle to a Game Object.
      *
@@ -3527,7 +3560,7 @@ export interface KAPLAYCtx {
      * as its own font.
      *
      * This waits for the sprite to load before doing anything, but if the sprite doesn't load, the game
-     * will transition to the error screen after a timeout (which is set by {@link KAPLAYOpt.loadTimeout}).
+     * will transition to the error screen after a timeout (which is set by {@link KAPLAYOpt["loadTimeout"]|KAPLAYOpt.loadTimeout}).
      *
      * @param sprite - The ID of the sprite to use as a font. Must already have frames defined
      * @param chars - The characters that correspond to each of the frames in the sprite. You can't use
@@ -3778,7 +3811,8 @@ export interface KAPLAYCtx {
      */
     dt(): number;
     /**
-     * Get the fixed delta time since last frame.
+     * Get the delta time for the fixed-update loop. This
+     * only changes when you call {@link setFixedSpeed}.
      *
      * @since v3000.0
      * @group Info
@@ -3791,6 +3825,14 @@ export interface KAPLAYCtx {
      * @group Info
      */
     restDt(): number;
+    /**
+     * Change the speed that the fixed update loop runs at.
+     * The options (and caveats) are described at {@link KAPLAYOpt["fixedUpdateMode"]|KAPLAYOpt.fixedUpdateMode}.
+     *
+     * @group Physics
+     * @experimental
+     */
+    setFixedSpeed(speed: FixedSpeedOption): void;
     /**
      * Get the total time since beginning.
      *
@@ -5307,6 +5349,23 @@ export interface KAPLAYCtx {
         continuity: number,
         bias: number,
     ): (t: number) => Vec2;
+    createRegularPolygon(
+        radius: number,
+        sides: number,
+        startAngle: number,
+    ): Vec2[];
+    createStarPolygon(
+        radius1: number,
+        radius2: number,
+        sides: number,
+        startAngle: number,
+    ): Vec2[];
+    createCogPolygon(
+        radius1: number,
+        radius2: number,
+        sides: number,
+        startAngle: number,
+    ): Vec2[];
     /**
      * Check if a line and a point intersect.
      *
@@ -5552,6 +5611,12 @@ export interface KAPLAYCtx {
      */
     Mat4: typeof Mat4;
     /**
+     * @since v3001.0
+     * @group Math
+     * @subgroup Advanced
+     */
+    Mat2: typeof Mat2;
+    /**
      * @since v4000.0
      * @group Math
      * @subgroup Advanced
@@ -5565,6 +5630,30 @@ export interface KAPLAYCtx {
      * @subgroup Advanced
      */
     Quad: typeof Quad;
+    /**
+     * A quadtree
+     *
+     * @since 4000
+     * @group Math
+     * @subgroup Advanced
+     */
+    Quadtree: typeof Quadtree;
+    /**
+     * Make a new quadtree
+     *
+     * @param pos - The position of the top level node
+     * @param width - The width of the top level node
+     * @param height - The height of the top level node
+     * @param maxObjects - The maximum amount of objects per node before splitting
+     * @param maxLevels - The maximum amount of levels
+     */
+    makeQuadtree(
+        pos: Vec2,
+        width: number,
+        height: number,
+        maxObjects: number,
+        maxLevels: number,
+    ): Quadtree;
     /**
      * The Random Number Generator.
      *