+++ src/ecs/components/transform/constraint.ts
@@ -1,6 +1,18 @@
 import { onAdd, onDestroy, onUnuse, onUse } from "../../../events/globalEvents";
+import { drawCircle } from "../../../gfx/draw/drawCircle";
+import { drawLine } from "../../../gfx/draw/drawLine";
+import { drawPolygon } from "../../../gfx/draw/drawPolygon";
+import {
+    loadMatrix,
+    multRotate,
+    popTransform,
+    pushMatrix,
+    pushTransform,
+} from "../../../gfx/stack";
+import { clamp } from "../../../math/clamp";
+import { Color } from "../../../math/color";
 import { lerp } from "../../../math/lerp";
-import { rad2deg, vec2 } from "../../../math/math";
+import { deg2rad, rad2deg, vec2 } from "../../../math/math";
 import {
     calcTransform,
     clampAngle,
@@ -488,7 +500,12 @@ export const constraint = {
             },
         };
     },
-    bone(minAngle?: number, maxAngle?: number) {
+    bone(
+        minAngle?: number,
+        maxAngle?: number,
+        minLength?: number,
+        maxLength?: number,
+    ) {
         let _minAngle = Math.max(
             -180,
             Math.min(minAngle ?? -180, maxAngle ?? 180),
@@ -497,6 +514,8 @@ export const constraint = {
             180,
             Math.max(minAngle ?? -180, maxAngle ?? 180),
         );
+        let _minLength = minLength;
+        let _maxLength = maxLength;
         return {
             id: "bone",
             get minAngle() {
@@ -514,6 +533,22 @@ export const constraint = {
                     180,
                     Math.max(minAngle ?? -180, maxAngle ?? 180),
                 );
+            },
+            get minLength() {
+                return _minLength;
+            },
+            get maxLength() {
+                return _maxLength;
+            },
+            drawInspect(this: GameObj<IKConstraintComp | RotateComp>) {
+                pushTransform();
+                multRotate(-this.angle + _minAngle);
+                drawCircle({
+                    radius: 20,
+                    start: _maxAngle - _minAngle,
+                    color: Color.RED,
+                });
+                popTransform();
             },
         };
     },
@@ -589,7 +624,10 @@ export const constraint = {
                                     transform.getRotation(),
                                 );
                                 // If constraint on angle, apply
-                                if (effector.minAngle && effector.maxAngle) {
+                                if (
+                                    effector.minAngle != undefined
+                                    && effector.maxAngle != undefined
+                                ) {
                                     newAngle = Math.min(
                                         Math.max(newAngle, effector.minAngle),
                                         effector.maxAngle,
@@ -603,7 +641,10 @@ export const constraint = {
                                     rotation + angleCorrection,
                                 );
                                 // If constraint on angle, apply
-                                if (effector.minAngle && effector.maxAngle) {
+                                if (
+                                    effector.minAngle != undefined
+                                    && effector.maxAngle != undefined
+                                ) {
                                     newAngle = Math.min(
                                         Math.max(newAngle, effector.minAngle),
                                         effector.maxAngle,
@@ -612,7 +653,50 @@ export const constraint = {
                                 effector.angle = newAngle;
                             }
 
-                            if (effector.minAngle && effector.maxAngle) {
+                            if (
+                                effector.parent
+                                && effector.parent.minLength != undefined
+                                && effector.parent.maxLength != undefined
+                            ) {
+                                // Vector from parent to target
+                                const ax = target.transform.e
+                                    - effector.parent.transform.e;
+                                const ay = target.transform.f
+                                    - effector.parent.transform.f;
+                                // Vector from parent to bone
+                                const bx = effectorTransform.e
+                                    - effector.parent.transform.e;
+                                const by = effectorTransform.f
+                                    - effector.parent.transform.f;
+                                // Projection of target onto bone
+                                const s = (ax * bx + ay * by)
+                                    / (bx * bx + by * by);
+                                const px = s * bx;
+                                const py = s * by;
+                                // Desired length
+                                let len = Math.sqrt(px * px + py * py);
+                                // New length = desired length clamped
+                                const nlen = clamp(
+                                    len,
+                                    effector.parent.minLength,
+                                    effector.parent.maxLength,
+                                );
+                                // Old length
+                                const olen = Math.sqrt(
+                                    effector.pos.x * effector.pos.x
+                                        + effector.pos.y * effector.pos.y,
+                                );
+                                // Scale to desired length clamped
+                                effector.pos.x *= nlen / olen;
+                                effector.pos.y *= nlen / olen;
+                            }
+
+                            if (
+                                (effector.minAngle != undefined
+                                    && effector.maxAngle != undefined)
+                                || (effector.minLength != undefined
+                                    && effector.maxLength != undefined)
+                            ) {
                                 // We changed the local angle, so the current effector's transform needs to be updated
                                 updateTransformRecursive(effector);
                             }
@@ -621,6 +705,29 @@ export const constraint = {
                             }
                         }
                     }
+                },
+                drawInspect(this: GameObj<IKConstraintComp>) {
+                    const endEffector = chain[0] = this;
+                    for (let i = 1; i <= depth; i++) {
+                        chain[i] = chain[i - 1].parent!;
+                    }
+                    let p1 = chain[depth].pos;
+                    pushTransform();
+                    for (let i = depth; i > 0; i--) {
+                        const bone = chain[i];
+                        const childBone = chain[i - 1];
+                        const len = childBone.pos.len();
+                        loadMatrix(bone.transform);
+                        drawPolygon({
+                            pts: [
+                                vec2(0, 0),
+                                vec2(len / 10, -len / 10),
+                                childBone.pos,
+                                vec2(len / 10, len / 10),
+                            ],
+                        });
+                    }
+                    popTransform();
                 },
             };
         }
@@ -744,6 +851,28 @@ export const constraint = {
                             obj.pos.y = obj.transform.f;
                         }
                     }
+                },
+                drawInspect(this: GameObj<IKConstraintComp>) {
+                    const endEffector = chain[0] = this;
+                    for (let i = 1; i <= depth; i++) {
+                        chain[i] = chain[i - 1].parent!;
+                    }
+                    let p1 = chain[depth].pos;
+                    pushTransform();
+                    for (let i = depth; i > 0; i--) {
+                        const bone = chain[i];
+                        const childBone = chain[i - 1];
+                        loadMatrix(bone.transform);
+                        drawPolygon({
+                            pts: [
+                                vec2(0, 0),
+                                vec2(10, -10),
+                                childBone.pos,
+                                vec2(10, 10),
+                            ],
+                        });
+                    }
+                    popTransform();
                 },
             };
         }