+++ src/gfx/formatText.ts
@@ -26,6 +26,7 @@ export type FontAtlas = {
     font: BitmapFontData;
     cursor: Vec2;
     maxHeight: number;
+    maxActualBoundingBoxAscent: number;
     outline: Outline | null;
 };
 
@@ -170,6 +171,7 @@ function getFontAtlasForFont(font: FontData | string): FontAtlas {
             },
             cursor: new Vec2(0),
             maxHeight: 0,
+            maxActualBoundingBoxAscent: 0,
             outline: opts.outline,
         };
 
@@ -177,6 +179,14 @@ function getFontAtlasForFont(font: FontData | string): FontAtlas {
     }
     return atlas;
 }
+
+const allChars = () => {
+    const renderableChars: string[] = [];
+    for (let i = 32; i <= 128; i++) { // Common Unicode range
+        renderableChars.push(String.fromCharCode(i));
+    }
+    return renderableChars.join("");
+};
 
 function updateFontAtlas(font: FontData | string, ch: string) {
     const atlas = getFontAtlasForFont(font);
@@ -201,11 +211,18 @@ function updateFontAtlas(font: FontData | string, ch: string) {
         c2d.textBaseline = "top";
         c2d.textAlign = "left";
         c2d.fillStyle = "#ffffff";
+
+        if (atlas.maxActualBoundingBoxAscent === 0) {
+            atlas.maxActualBoundingBoxAscent =
+                c2d.measureText(allChars()).actualBoundingBoxAscent;
+        }
+        const maxActualBoundingBoxAscent = atlas.maxActualBoundingBoxAscent;
         const m = c2d.measureText(ch);
         let w = Math.ceil(m.width);
         if (!w) return;
-        let h = (Math.ceil(Math.abs(m.actualBoundingBoxAscent))
-            + Math.ceil(Math.abs(m.actualBoundingBoxDescent)))
+        let h = maxActualBoundingBoxAscent
+                + Math.ceil(Math.abs(m.actualBoundingBoxAscent))
+                + Math.ceil(Math.abs(m.actualBoundingBoxDescent))
             || atlas.font.size;
 
         // TODO: Test if this works with the verification of width and color
@@ -229,7 +246,7 @@ function updateFontAtlas(font: FontData | string, ch: string) {
         c2d.fillText(
             ch,
             atlas.outline?.width ?? 0,
-            atlas.outline?.width ?? 0,
+            (atlas.outline?.width ?? 0) + maxActualBoundingBoxAscent,
         );
 
         const img = c2d.getImageData(
@@ -258,7 +275,7 @@ function updateFontAtlas(font: FontData | string, ch: string) {
             atlas.cursor.x,
             atlas.cursor.y,
             w,
-            h,
+            h + maxActualBoundingBoxAscent,
         );
 
         atlas.cursor.x += w + 1;