+++ CHANGELOG.md
@@ -18,11 +18,36 @@ best friend, lajbel, can put the correct version name here
 
 ## [unreleased]
 
+## [4000.0.0-alpha.24] - 2025-12-12
+
 ### Added
 
+- Added the `maxTimeStep` and `fixedUpdateMode` options, as well as
+  `setFixedSpeed()` for more granular control over fixed update and timing -
+  @dragoncoder047
 - Added parameterized formatting tags like `"[color=red]Red text![/color]"` in
   `CharTransformFunc` for more powerful text formatting options -
   @dragoncoder047
+- Added `createRegularPolygon()` and `createStarPolygon()` to create 2D regular
+  polytopes - @mflerackers
+- Added `createCogPolygon()` to create 2D regular cogs - @mflerackers
+- Added `getSpriteOutline()` that takes a sprite asset and returns a polygon
+  showing the outline - @milosilo-dev
+- Added Quadtree for collision detection (only for fixed size screen for now,
+  needs expansion) - @mflerackers
+- Added vertical sweep and prune - @mflerackers
+- Added configuration to choose broad phase algorithm - @mflerackers
+
+### Fixed
+
+- Fixed the `fakeMouse()` component not giving the right position when the
+  camera transform was not the identity matrix - @dragoncoder047
+- Fixed tall fonts being cropped - @anthonygood
+- Fixed the sprite animation `onEnd()` callback being called before the
+  animation actually stopped, so if the onEnd callback started a new animation,
+  the new animation was instantly stopped - @dragoncoder047
+- Now `playMusic()` actually uses the requested volume and playback rate given
+  in the options - @dragoncoder047
 
 ## [4000.0.0-alpha.23] - 2025-11-05
 
@@ -73,7 +98,8 @@ best friend, lajbel, can put the correct version name here
           // runs until obj is destroyed
       });
 
-      scene.onUpdate(() => { // or just onUpdate(() => {
+      scene.onUpdate(() => {
+          // or just onUpdate(() => {
           // runs until scene is changed
       });
   });
@@ -114,6 +140,7 @@ best friend, lajbel, can put the correct version name here
   - `onDraw()`
 
 - Added `app` scope for app event handlers - @lajbel
+
   ```js
   app.onUpdate(() => {
       // runs until it is cancelled
@@ -145,7 +172,8 @@ best friend, lajbel, can put the correct version name here
 
   ```js
   scene("game", () => {
-      scene.onUpdate(() => { // or just onUpdate(() => {
+      scene.onUpdate(() => {
+          // or just onUpdate(() => {
           // runs until scene is changed
       });
   });
@@ -263,10 +291,7 @@ best friend, lajbel, can put the correct version name here
   @lajbel
   ```js
   // blue frog
-  add([
-      sprite("bean"),
-      color(0x0000ff),
-  ]);
+  add([sprite("bean"), color(0x0000ff)]);
   ```
 - **(!)** `KAPLAYCtx` doesn't use generics anymore. Now, `KAPLAYCtxT` uses
   them - @lajbel